Ultime modifiche di Line e Stop

+++ src/main/java/com/generation/bus/repository/LineRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.generation.bus.model.Line;
 
-public interface LineRepository extends JpaRepository<Line, Integer> {
+public interface LineRepository extends JpaRepository<Line, Long> {
 
 }
